+++ codegen/src/test/resources/SignalArgShapesTest/interface.1.kt
@@ -1,0 +1,11 @@
+package org.example.signals
+
+import com.monkopedia.sdbus.InterfaceName
+
+public interface Shapes {
+  public fun register()
+
+  public companion object {
+    public val INTERFACE_NAME: InterfaceName = InterfaceName("org.example.signals.Shapes")
+  }
+}